requirements.txt
ipython
python-gitlab
fs.dropboxfs
fs.onedrivefs
google-auth-httplib2
google-auth-oauthlib
